src/requirements.txt
a2a-sdk==0.2.*
git+https://github.com/google/adk-python.git@main
streamlit
tabulate
platformdirs
Flask
gunicorn
scikit-learn
google-genai==1.19.*
pillow
pandas
google-cloud-aiplatform[adk]==1.97.0
google-cloud-bigquery[pandas]
google-cloud-bigquery-storage
google-cloud-firestore==2.21.*
altair==5.5.*
vl-convert-python
python-dotenv
cloudpickle==3.1.1
pydantic==2.11.*
yfinance
matplotlib
